System components refactoring
Trying to fix up systems components to match how other components are
specified. For example, instead of including the same pump in every
system, have a pump definition defined at the project level, then have
references to that definition in the system, like how plug loads in
spaces work. Also, there appears to be no need for an auditMotor, as it
is a secondary component of pumps and fans and does not need its own
specification.